# 00-PROJECTS/HAIDA/haida-main/dev/docs/API/openapi.yaml
openapi: 3.0.3
info:
  title: HAIDA API
  description: |
    HAIDA - Quality Assurance Platform with AI Integration

    A comprehensive testing and auditing platform that combines traditional QA automation
    with AI-powered insights for enterprise software quality assurance.

    ## Key Features
    - Multi-tenant architecture with RBAC
    - Automated testing execution and monitoring
    - AI-powered chat and assistance (Copilot Studio)
    - Comprehensive documentation management
    - Real-time analytics and reporting
    - Enterprise SSO integration (Microsoft Entra ID)
  version: 2.0.0
  contact:
    name: HAIDA Development Team
    email: hola@stayarta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://haida.vercel.app/api/v2
    description: Production server
  - url: http://localhost:8000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - entraAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    entraAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes:
            User.Read: Read user profile
            Files.Read.All: Read files
            Calendars.Read: Read calendar
            Mail.Read: Read mail
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2025-12-16T10:30:00Z"
...
    Report:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        tenant_id:
          $ref: '#/components/schemas/UUID'
        template_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        report_type:
          type: string
          enum: [test_execution, coverage, performance, custom]
        status:
          type: string
          enum: [generating, completed, failed]
        format:
          type: string
          enum: [pdf, html, json, xlsx]
          default: pdf
        file_url:
          type: string
        generated_at:
          $ref: '#/components/schemas/Timestamp'
  parameters:
    tenantId:
      name: tenant_id
      in: header
      description: Tenant/workspace identifier (UUID or slug)
      required: false
      schema:
        type: string
    page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
paths:
...
  /reports:
    get:
      summary: List reports
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: report_type
          in: query
          schema:
            type: string
            enum: [test_execution, coverage, performance, custom]
      responses:
        '200':
          description: Reports retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'

    post:
      summary: Generate report
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, report_type]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                description:
                  type: string
                report_type:
                  type: string
                  enum: [test_execution, coverage, performance, custom]
                template_id:
                  $ref: '#/components/schemas/UUID'
                parameters:
                  type: object
                format:
                  type: string
                  enum: [pdf, html, json, xlsx]
                  default: pdf
      responses:
        '202':
          description: Report generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    $ref: '#/components/schemas/UUID'
                  status:
                    type: string
                    example: generating
                  estimated_completion:
                    $ref: '#/components/schemas/Timestamp'
  /reports/{reportId}/download:
    get:
      summary: Download report
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Report file download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Report not found or not ready

  # System & Observability
...